[MOD] XQuery: Implementation of arrays as Hinze and Patterson's Fingertrees.

+++ basex-core/src/main/java/org/basex/query/func/array/ArrayFn.java
@@ -21,10 +21,11 @@ abstract class ArrayFn extends StandardFunc {
    * @return specified position -1
    * @throws QueryException query exception
    */
-  final int checkPos(final Array array, final long pos, final boolean incl) throws QueryException {
-    final int as = array.arraySize() + (incl ? 1 : 0);
+  final long checkPos(final Array array, final long pos, final boolean incl)
+      throws QueryException {
+    final long as = array.arraySize() + (incl ? 1 : 0);
     if(pos < 1 || pos > as) throw (as == 0 ? ARRAYEMPTY : ARRAYBOUNDS_X_X).get(info, pos, as);
-    return (int) pos - 1;
+    return pos - 1;
   }
 
   /**
@@ -34,7 +35,7 @@ abstract class ArrayFn extends StandardFunc {
    * @return specified position -1
    * @throws QueryException query exception
    */
-  final int checkPos(final Array array, final long pos) throws QueryException {
+  final long checkPos(final Array array, final long pos) throws QueryException {
     return checkPos(array, pos, false);
   }
 }
